Worked on the Chat
Hello everyone :night_with_stars: ,

I tried to work on the Chat but it's kind of hard to put together with a GIU so for now stop the dev for it.

I will continue later , to calm down and stop to rage over it LOL.

By the way I also made some minors changes on other files.

Thank You,
Franck :hearts: