MAP Time
Hello Everyone :smile: ,

Today I want to give you a new update with a cool feature: A dynamic MAP on Tkinter .

So this map will be updated when you use the feature and a flag on it will appear to tell you where you are on it .
We are using the YANDEX Static MAP Service for this feature.

Very classic but very useful.

I hope that you will enjoy this new update! :heart:

Good night :bridge_at_night:

Franck :+1: .